terraform/modules/common/any/payload/linux/modules/deploy-cloudsql-app/resources/requirements.txt
Flask==2.1.3
Flask_API==3.0.post1
pymysql==1.0.3
google-auth==2.26.2